Group test cases instead of defining just valid/invalid cases
